% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcs_footer.R
\name{bcsFooterUI}
\alias{bcsFooterUI}
\title{Module UI to for BCStats footer}
\usage{
bcsFooterUI(id)
}
\arguments{
\item{id}{character}
}
\value{
a UI element
}
\description{
Module UI to for BCStats footer
}
