% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcs_header.R
\name{bcsHeaderServer}
\alias{bcsHeaderServer}
\title{Server-side for BC Stats header}
\usage{
bcsHeaderServer(id, links = TRUE)
}
\arguments{
\item{id}{character}

\item{links}{logical}
}
\description{
Server-side for BC Stats header
}
