% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_bcstats_app}
\alias{create_bcstats_app}
\title{Create an app project}
\usage{
create_bcstats_app(
  path = ".",
  coc_email = bcgovr:::get_coc_email(),
  licence = "apache2"
)
}
\arguments{
\item{path}{Path to the directory in which to initialize the project. Default "." - your current working directory.}

\item{coc_email}{Contact email address(es) for the Code of Conduct.}

\item{licence}{Which licence to apply? Default is Apache 2.0 ("apache2"). Use "cc-by" for Creative Commons Attribution 4.0}
}
\description{
Create an app project in the BCStats style, providing all the files needed to get started
}
