% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcs_header.R
\name{bcsHeaderUI}
\alias{bcsHeaderUI}
\title{Module UI to for BCStats header}
\usage{
bcsHeaderUI(id, appname, github = NULL)
}
\arguments{
\item{id}{character}

\item{appname}{character}

\item{github}{character url to github repo or NULL}
}
\value{
a UI element
}
\description{
Module UI to for BCStats header
}
